Remove the octets numbers from the chunked response for blueprint output
Chunked transfer encoding format includes the number of the octets
of the data in each chunk as a part of the response.

This commit adds detection of the `Transfer-Encoding: chunked`
into the parser to strip the numbers from the output for blueprint
format. The output of raw format kept not modified.

This fix probably also resolves issue apiaryio/curl-trace-parser#6.